refactor: minimize changes and add settings.ts
