+++ lib/application-stack.ts
